Update configs and init step
All config fields are now optional, since the user might need to run only a subset of pipeline steps. The init subcommand can now produce a simplified default config, a config to run the example data, or an extended config with all fields. It can now reproduce the snakemake and shell example directories.