Normalize using read depth
Previously, coverage was normalized by cumulative coverage per gene, which addressed differing read depths, but also removed total expression information. Now, in the setup stage, total coverage across a random sample of 1000 genes is used to compute sample-specific scaling factors relative to the median. These are used in the coverage stage for normalization. If applying existing models to new samples, the setup stage loads the median total coverage value and list of genes used in the previous setup, and uses those to compute the scaling factors.