Scaling factors from bigWig sumData
Previously, a random sample of genes was chosen, and coverage sum across those genes was used to estimate per-sample coverage scaling factors to use for getting binned coverage. Since bigWig files have a `sumData` field in their header, that is used first to calculate the median genome-wide coverage sum during setup, which is saved in `info/`. Then, the coverage step uses that and the sumData field for each bigWig it loads to scale each value to control for sequencing depth.

This scaling procedure is now also used during setup for estimating the variance threshold for the relevant adaptive binning methods. As in the coverage step, the per-base values are scaled before log-transformation to reduce the effect of variable sequencing depths.