Snakefile resources, bigWig paths
Added time and memory resources to the more intensive snakemake rules. Fixes #5

Added the option to specify the root bigWig directory in the config, and the paths in the manifest file will be relative to that. This reduces redundancy in the manifest file and makes it easier to move the bigWig location without having to edit the manifest.