Ignore gene ID version in GTF
The first period and all subsequent characters in a gene ID are assumed to be the version (e.g. for Ensembl IDs) and are stripped off in the setup stage and excluded from output files.

- It may make sense for gene annotations to include ID versions since they give precise coordinates, but the latent phenotypes don't need to be specified so specifically.
- There is potential difficulty or confusion when running this in "residual phenotype" mode, if the loaded explicit phenotypes don't include version numbers, or the version numbers differ, in which case they won't be regressed out.

ID versions are also ignored in the loaded explicit phenotype tables to prevent mismatches for the same ID. And while version numbers should no longer be the source of the code silently failing to regress out phenotypes due to ID mismatches, it could happen for another reason. So the coverage stage now reports how many of the genes in a batch had explicit phenotypes to regress out.