Add gene filtering and lncRNA
Genes for which too many samples (>50% by default) have zero coverage are excluded in the fit and transform stages. This is because their models and phenotypes will be noisy at best, and if a phenotype has too many of the same values across samples, genetic mapping could produce spurious associations. Fixes #18

This filtering is especially important for lncRNA, because there can be many annotated lncRNA with little or no expression in the tissue. The setup stage now includes genes of type "lncRNA" in addition to "protein_coding".